Remove pom_test_runner.py: Eliminated the POM + Test Runner script to streamline the project structure and focus on core functionalities.
